requirements.txt
alembic==1.4.2
amqp==2.6.0
aniso8601==8.0.0
apiwrapper==0.1.8
APScheduler==3.6.3
arrow==0.15.8
astroid==2.4.2
autopep8==1.5.3
billiard==3.6.3.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
croniter==0.3.34
cryptography==3.0
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Admin==1.5.6
Flask-APScheduler==1.11.0
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.8
flask-crontab==0.1.2
Flask-Inputs==0.3.0
Flask-JSON==0.3.4
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-RQ2==18.3
Flask-SQLAlchemy==2.4.4
future==0.18.2
geojson==2.5.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflect==4.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
mysqlclient==2.0.1
natsort==7.0.1
ndg-httpsclient==0.5.1
numpy==1.19.1
oauthlib==3.1.0
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
pylint==2.5.3
PyMySQL==0.10.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-telegram-bot==12.8
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rq==1.5.0
rq-dashboard==0.6.1
rq-scheduler==0.10.0
sendotp==0.0.4
sentry-sdk==0.16.3
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
tornado==6.0.4
tweepy==3.9.0
twilio==6.44.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.10
vine==1.3.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.1.0
